requirements.txt
python-docx
pdfplumber
streamlit
fpdf
google.generativeai
